requirements.txt
Django==4.0.4
djangorestframework==3.13.1
emoji==1.7.0
environs==8.0.0
loguru==0.6.0
Pillow==9.1.0
